Support dynamic rebinding of listen address
This patch allows your Redis to rebind to a new IP address, or unbind
the TCP socket altogether.

In the config file, use:

  bind -

to tell Redis to not bind any TCP sockets. You should use this instead
of setting port to 0 so that the socket may be rebound at a later time.
Note that you must set "unixsocket" so that you can access Redis in this
state.

To dynamically rebind Redis' listening socket, send Redis a CONFIG SET
BIND command, e.g.:

  CONFIG SET BIND 192.0.2.42