New command: pipesave
Allows you to specify a command to run (via the config file *only*, to
prevent people from running arbitrary code on your machine via redis), which
will be invoked and fed the contents of an RDB dump file.

Useful for taking offsite backups (say, to S3 or a centralised backup server)
or a periodic replication mechanism when you don't want (or don't have the
available memory) to run twice as many redis instances everywhere.

Redis treats these children as it would treat a child forked to BGSAVE.
This simplifies the code (rdb_child_pid is already checked where it matters),
but means you can't run a BGSAVE and PIPESAVE concurrently.

Conflicts:
	src/config.h