Change the way we handle overmemory conditions in NDS, esp. on RDB load
Turns out that letting Redis balloon out during RDB load is an absolutely terrible idea.
Firstly, it consumes a lot of memory, which is explicitly what we're trying to avoid,
and it also results in *massive* MDB flushes, which MDB is not great friends with.

Instead, what we'll do now is return REDIS_ERR if we can't free memory (which is the
original behaviour), and handle that specially in the one place where we *don't* want
to provide a user-visible error: during an RDB load (replication initiation).  Since
we don't want to endlessly loop if our RDB load can't manage memory properly, we also
check if there's problems and exit from the load if necessary.